+++ lib/src/directives/drag.directive.ts
@@ -1,45 +1,53 @@
-import { AfterViewInit, Directive, ElementRef, Input, NgZone, OnDestroy, Renderer2 } from '@angular/core';
+import { AfterViewInit, Directive, ElementRef, Input, NgZone, OnDestroy, Renderer2, Output, EventEmitter } from '@angular/core';
 import { Subject, fromEvent } from 'rxjs';
-import { takeUntil, map, switchMap } from 'rxjs/operators';
+import { takeUntil, map, switchMap, filter } from 'rxjs/operators';
+import { isString } from '@datorama/utils';
+import { setStyle } from '../internal/helpers';
+
+export type DraggedEvent = {
+  x: number;
+  y: number;
+};
 
 @Directive({
   selector: '[datoDraggable]'
 })
 export class DatoDraggableDirective implements AfterViewInit, OnDestroy {
-  @Input() datoDragHandle: string | Element;
-  @Input() datoDragTarget: string | Element;
-  @Input() datoDragEnabled: boolean = true;
+  @Input()
+  datoDragHandle: string | Element;
+  @Input()
+  datoDragTarget: string | Element;
+  @Input()
+  set datoDragEnabled(enabled) {
+    this.enabled = enabled;
+    /** determine if the component has been init by the handle variable */
+    if (this.handle) {
+      setStyle(this.handle, 'cursor', enabled ? 'move' : 'default');
+    } else if (enabled) {
+      this.init();
+    }
+  }
+  @Output()
+  dragged = new EventEmitter<DraggedEvent>();
 
-  // Element to be dragged
+  /** Element to be dragged */
   private target: Element;
-  // Drag handle
+  /** Drag handle */
   private handle: Element;
   private delta = { x: 0, y: 0 };
   private offset = { x: 0, y: 0 };
-
+  private enabled = true;
   private destroy$ = new Subject<void>();
+  private enabledFilter = source => source.pipe(filter(() => this.enabled));
 
   constructor(private host: ElementRef, private zone: NgZone, private renderer: Renderer2) {}
 
   public ngAfterViewInit(): void {
-    if (!this.datoDragEnabled) {
+    if (!this.enabled) {
       return;
     }
 
-    if (!this.datoDragTarget) {
-      throw 'You need to specify the drag target';
-    }
-
-    this.handle = this.datoDragHandle instanceof Element ? this.datoDragHandle : typeof this.datoDragHandle === 'string' && this.datoDragHandle ? document.querySelector(this.datoDragHandle as string) : this.host.nativeElement;
-
-    // add the move cursor
-    if (this.handle) {
-      this.renderer.setStyle(this.handle, 'cursor', 'move');
-    }
-
-    this.target = this.datoDragTarget instanceof Element ? this.datoDragTarget : document.querySelector(this.datoDragTarget as string);
-
-    this.setupEvents();
+    this.init();
   }
 
   public ngOnDestroy(): void {
@@ -52,25 +60,25 @@ export class DatoDraggableDirective implements AfterViewInit, OnDestroy {
       let mousemove$ = fromEvent(document, 'mousemove');
       let mouseup$ = fromEvent(document, 'mouseup');
 
-      let mousedrag$ = mousedown$
-        .pipe(
-          switchMap((event: MouseEvent) => {
-            let startX = event.clientX;
-            let startY = event.clientY;
+      let mousedrag$ = mousedown$.pipe(
+        this.enabledFilter,
+        switchMap((event: MouseEvent) => {
+          let startX = event.clientX;
+          let startY = event.clientY;
 
-            return mousemove$.pipe(
-              map((event: MouseEvent) => {
-                event.preventDefault();
-                this.delta = {
-                  x: event.clientX - startX,
-                  y: event.clientY - startY
-                };
-              }),
-              takeUntil(mouseup$)
-            );
-          })
-        )
-        .pipe(takeUntil(this.destroy$));
+          return mousemove$.pipe(
+            map((event: MouseEvent) => {
+              event.preventDefault();
+              this.delta = {
+                x: event.clientX - startX,
+                y: event.clientY - startY
+              };
+            }),
+            takeUntil(mouseup$)
+          );
+        }),
+        takeUntil(this.destroy$)
+      );
 
       mousedrag$.subscribe(() => {
         if (this.delta.x === 0 && this.delta.y === 0) {
@@ -80,11 +88,17 @@ export class DatoDraggableDirective implements AfterViewInit, OnDestroy {
         this.translate();
       });
 
-      mouseup$.pipe(takeUntil(this.destroy$)).subscribe(() => {
-        this.offset.x += this.delta.x;
-        this.offset.y += this.delta.y;
-        this.delta = { x: 0, y: 0 };
-      });
+      mouseup$
+        .pipe(
+          this.enabledFilter,
+          takeUntil(this.destroy$)
+        )
+        .subscribe(() => {
+          this.offset.x += this.delta.x;
+          this.offset.y += this.delta.y;
+          this.dragged.emit(this.offset);
+          this.delta = { x: 0, y: 0 };
+        });
     });
   }
 
@@ -95,5 +109,25 @@ export class DatoDraggableDirective implements AfterViewInit, OnDestroy {
                   ${this.offset.y + this.delta.y}px)
       `;
     });
+  }
+
+  /**
+   * Init the directive
+   */
+  private init() {
+    if (!this.datoDragTarget) {
+      throw new Error('You need to specify the drag target');
+    }
+
+    this.handle = this.datoDragHandle instanceof Element ? this.datoDragHandle : isString(this.datoDragHandle) && this.datoDragHandle ? document.querySelector(this.datoDragHandle as string) : this.host.nativeElement;
+
+    /** add the move cursor */
+    if (this.handle && this.enabled) {
+      setStyle(this.handle, 'cursor', 'move');
+    }
+
+    this.target = this.datoDragTarget instanceof Element ? this.datoDragTarget : document.querySelector(this.datoDragTarget as string);
+
+    this.setupEvents();
   }
 }