+++ lib/src/rx/public_api.ts
@@ -9,3 +9,4 @@
 export * from './polling';
 export * from './debounce';
 export * from './worker';
+export * from './dirty-check';